feat [DEV-27]: Move media guards to separate folders
