+++ libs/angular/ui/src/modules/media/guards/index.ts
@@ -1,6 +1,6 @@
-export { mediaMinGuard } from './media-min.guard';
-export { mediaBetweenGuard } from './media-between.guard';
-export { mediaMaxGuard } from './media-max.guard';
-export { mediaOnlyGuard } from './media-only.guard';
-export { mediaDesktopGuard } from './media-desktop.guard';
-export { mediaMobileGuard } from './media-mobile.guard';
+export { mediaMinGuard } from './media-min';
+export { mediaBetweenGuard } from './media-between';
+export { mediaMaxGuard } from './media-max';
+export { mediaOnlyGuard } from './media-only';
+export { mediaDesktopGuard } from './media-desktop';
+export { mediaMobileGuard } from './media-mobile';
