+++ libs/angular/ui/src/modules/media/guards/media-between/index.ts
@@ -1,0 +1,1 @@
+export { mediaBetweenGuard } from './media-between.guard';
