+++ libs/angular/ui/src/modules/media/guards/media-desktop/index.ts
@@ -1,0 +1,1 @@
+export { mediaDesktopGuard } from './media-desktop.guard';
