+++ libs/angular/ui/src/modules/media/guards/media-max/index.ts
@@ -1,0 +1,1 @@
+export { mediaMaxGuard } from './media-max.guard';
