+++ libs/angular/ui/src/modules/media/guards/media-min/index.ts
@@ -1,0 +1,1 @@
+export { mediaMinGuard } from './media-min.guard';
