+++ libs/angular/ui/src/modules/media/guards/media-mobile/index.ts
@@ -1,0 +1,1 @@
+export { mediaMobileGuard } from './media-mobile.guard';
