+++ libs/angular/ui/src/modules/media/guards/media-only/index.ts
@@ -1,0 +1,1 @@
+export { mediaOnlyGuard } from './media-only.guard';
