+++ libs/angular/ui/src/modules/media/guards/media-only/media-only.guard.ts
@@ -2,8 +2,8 @@ import { inject } from '@angular/core';
 import { toSignal } from '@angular/core/rxjs-interop';
 import { CanActivateFn } from '@angular/router';
 
-import { MediaService } from '../services';
-import { MediaBreakpoint } from '../interfaces';
+import { MediaService } from '../../services';
+import { MediaBreakpoint } from '../../interfaces';
 
 /**
  * Guard to activate only on a specific media breakpoint.
